have a page that display clientlist and when clicking on client it can show recipes and restaurant for that client
